Exp-1/output/f00073_from_pretrained_test.py
from f00073_from_pretrained import *
def test_from_pretrained():
    # Test case 1
    model_name_or_path = 'distilbert-base-uncased'
    model = from_pretrained(model_name_or_path)
    assert isinstance(model, PreTrainedModel)
    
    # Test case 2
    model_name_or_path = 'path/to/model'
    model = from_pretrained(model_name_or_path)
    assert isinstance(model, PreTrainedModel)
    
    # Test case 3
    model_name_or_path = 'https://example.com/model'
    model = from_pretrained(model_name_or_path)
    assert isinstance(model, PreTrainedModel)
    
    # Test case 4
    model_name_or_path = 'invalid_path'
    model = from_pretrained(model_name_or_path)
    assert isinstance(model, PreTrainedModel)
    
    # Test case 5
    model_name_or_path = 'invalid_url'
    model = from_pretrained(model_name_or_path)
    assert isinstance(model, PreTrainedModel)
    
    print('All test cases passed!')
    

test_from_pretrained()
